+++ src/internal/RedBlackTree.ts
@@ -150,18 +150,19 @@ export const find = Dual.dual<
   <K>(key: K) => <V>(self: RBT.RedBlackTree<K, V>) => Chunk.Chunk<V>,
   <K, V>(self: RBT.RedBlackTree<K, V>, key: K) => Chunk.Chunk<V>
 >(2, <K, V>(self: RBT.RedBlackTree<K, V>, key: K) => {
-  const cmp = (self as RedBlackTreeImpl<K, V>)._ord.compare
+  const stack: Array<Node.Node<K, V>> = []
   let node = (self as RedBlackTreeImpl<K, V>)._root
   let result = Chunk.empty<V>()
-  while (node !== undefined) {
-    const d = cmp(key, node.key)
-    if (d === 0 && Equal.equals(key, node.key)) {
-      result = Chunk.prepend(node.value)(result)
-    }
-    if (d <= 0) {
+  while (node !== undefined || stack.length > 0) {
+    if (node) {
+      stack.push(node)
       node = node.left
     } else {
-      node = node.right
+      const current = stack.pop()!
+      if (Equal.equals(key, current.key)) {
+        result = Chunk.prepend(current.value)(result)
+      }
+      node = current.right
     }
   }
   return result